+++ changelog.md
@@ -23,6 +23,7 @@ Codion Change Log
 - DefaultEntityComboBoxModel bug fixed, now uses the comparator from the entity it is based on.
 - EntityComboBoxModel.ForeignKeyFilter.set() overloaded for a single Key.
 - EntityComboBoxModel.Builder.comparator() added.
+- DefaultEntityComboBoxModel.SelectedItemTranslator removed.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel now disposes all windows on exit.
 - EntityTablePanel, row limit configuration field now selects all on focus gained.